Update README.md
I tried to revise and condense the lecture while sticking to the original notes and structure.  The changes I made are reflective of the experience my colleague and I had teaching the content.  I also removed 'puts' from inside of methods except to illustrate when it is used it can return a value of NIL.  I instead added a brief explanation of using 'puts' when calling a method to print out the return value of the method in terminal.  This explanation distinguishes between the return value of a method and printing a method in terminal.